Ensure that project compiles on Scala 3
by introducing a Scala 3 stub implementation of
`erasedOptionalPrimitives`. Tests still fail
