+++ .github/workflows/ci.yml
@@ -59,8 +59,12 @@ jobs:
         run: sbt '++${{ matrix.scala }}' coverage test coverageReport
 
       - name: Scala build
-        if: '!startsWith(matrix.scala, ''2.13'')'
+        if: '!startsWith(matrix.scala, ''2.13'') && !startsWith(matrix.scala, ''3.0'')'
         run: sbt '++${{ matrix.scala }}' test
+
+      - name: Scala compile
+        if: startsWith(matrix.scala, '3.0')
+        run: sbt '++${{ matrix.scala }}' compile
 
       - name: Publish to Codecov.io
         if: startsWith(matrix.scala, '2.13')
