+++ src/main/scala-3/com/github/swagger/scala/converter/ErasureHelper.scala
@@ -1,0 +1,11 @@
+package com.github.swagger.scala.converter
+
+import io.swagger.v3.oas.models.media.Schema
+
+
+object ErasureHelper {
+
+  def erasedOptionalPrimitives(cls: Class[_]): Map[String, Class[_]] =  Map.empty[String, Class[_]]
+
+}
+
