+++ src/main/scala/com/github/swagger/scala/converter/SwaggerScalaModelConverter.scala
@@ -16,7 +16,6 @@ import io.swagger.v3.oas.models.media.Schema
 import org.slf4j.LoggerFactory
 
 import java.util
-import scala.reflect.runtime.universe
 import scala.util.Try
 import scala.util.control.NonFatal
 
@@ -38,30 +37,6 @@ class SwaggerScalaModelConverter extends ModelResolver(SwaggerScalaModelConverte
   private val BigIntClass = classOf[BigInt]
   private val ProductClass = classOf[Product]
   private val AnyClass = classOf[Any]
-
-  private def erasedOptionalPrimitives(cls: Class[_]): Map[String, Schema[_]] = {
-    val mirror = universe.runtimeMirror(cls.getClassLoader)
-    val sym = mirror.staticClass(cls.getName)
-    val properties = sym.selfType.members
-      .filterNot(_.isMethod)
-      .filterNot(_.isClass)
-      .map(prop => prop.name.toString.trim -> prop.typeSignature).toMap
-
-    properties.mapValues { typeSignature =>
-      if (mirror.runtimeClass(typeSignature.typeSymbol.asClass) != OptionClass) {
-        None
-      } else {
-        val typeArg = typeSignature.typeArgs.headOption
-        typeArg.flatMap { signature =>
-          if (signature.typeSymbol.isClass) {
-            val runtimeClass = mirror.runtimeClass(signature.typeSymbol.asClass)
-            Option(PrimitiveType.fromType(runtimeClass)).map(_.createProperty())
-          } else None
-        }
-      }
-    }.collect { case (k, Some(v)) => k -> v }.toMap
-  }
-
 
   override def resolve(`type`: AnnotatedType, context: ModelConverterContext, chain: Iterator[ModelConverter]): Schema[_] = {
     val javaType = _mapper.constructType(`type`.getType)
@@ -98,7 +73,7 @@ class SwaggerScalaModelConverter extends ModelResolver(SwaggerScalaModelConverte
 
   private def caseClassSchema(cls: Class[_], `type`: AnnotatedType, context: ModelConverterContext,
                               chain: util.Iterator[ModelConverter]): Option[Schema[_]] = {
-    val erasedProperties = erasedOptionalPrimitives(cls)
+    val erasedProperties = ErasureHelper.erasedOptionalPrimitives(cls)
 
     if (chain.hasNext) {
       Option(chain.next().resolve(`type`, context, chain)).map { schema =>
